+++ app/Main.hs
@@ -1,22 +1,23 @@
-module Main (main) where
+module Main where
 
-import System.IO
-import qualified Data.Text.IO as TIO
 import System.Environment (getArgs)
-
+import Client (startClient,handleGame)
+import Server (startServer)  -- 假设这是您的服务器启动函数
 import Logic(BoardDotStat,boardDotStatCons,
              WholeBoard,initWholeBoard,getDotStat,initWholeState,
              moveCursor,iPlacePawn,placePawnAtCursor,enemyPlacePawn,fetchWholeState)
-
-import Server(startServer)
-import Client(startClient)
 
 main :: IO ()
 main = do
     args <- getArgs
     case args of
-        ["server"] -> startServer
-        -- ["client", host, port] -> startClient host port
-        ["client"] -> startClient "127.0.0.1" "3000"
-        -- ["client"] -> startClient "https://super-duper-orbit-4vpvv7xw7pgcq4p6-3000.app.github.dev/" "3000"
-        _ -> putStrLn "Invalid arguments"
+        -- 启动客户端
+        ["client"] -> do
+            putStrLn "Starting client..."
+            handle <- startClient "127.0.0.1" "3000"
+            handleGame handle
+        -- 启动服务器
+        ["server"] -> do
+            putStrLn "Starting server..."
+            startServer
+        _ -> putStrLn "Usage: program (server | client <host> <port>)"
