+++ src/Client.hs
@@ -1,18 +1,15 @@
-module Client(startClient) where
+module Client (startClient, placePawn, getEnemyPawn, handleGame) where
 
 import Network.Socket
 import System.IO
-import Control.Monad (when, unless)
+import Control.Monad (unless, when)
+import Logic (WholeState, placePawn, iPlacePawn, enemyPlacePawn, initWholeState,fetchTurn)
 import Data.List.Split (splitOn)
-
-import Logic(BoardDotStat(..),boardDotStatCons,
-             WholeBoard(..),initWholeBoard,getDotStat,WholeState(..),initWholeState,getCorrespondentPawn,
-             moveCursor,iPlacePawn,placePawnAtCursor,enemyPlacePawn,placePawn,fetchWholeState, PlayerSide, WinStat)
-import Data.List (isPrefixOf)
 
 type Coordinate = (Int, Int)
 
-startClient :: String -> String -> IO ()
+-- 连接到服务器
+startClient :: String -> String -> IO Handle
 startClient host port = withSocketsDo $ do
     addrinfos <- getAddrInfo Nothing (Just host) (Just port)
     let serveraddr = head addrinfos
@@ -20,92 +17,70 @@ startClient host port = withSocketsDo $ do
     connect sock (addrAddress serveraddr)
     handle <- socketToHandle sock ReadWriteMode
     hSetBuffering handle NoBuffering
-    interactWithServer handle
-    hClose handle
+    return handle
 
-interactWithServer :: Handle -> IO ()
-interactWithServer handle = do
-    putStrLn "Connected to server. Type 'quit' to exit."
-    handleServerMessages handle
-    where
-        handleServerMessages handle = do
-            -- 读取并打印服务器响应
-            serverResponse <- hGetLine handle
-            putStrLn $ "Server says: " ++ serverResponse
+-- 发送坐标到服务器
+sendMyPawn :: Coordinate -> Handle -> IO ()
+sendMyPawn (x, y) handle = do
+    hPutStrLn handle (show x ++ "," ++ show y)
 
-            -- 根据服务器请求的不同响应相应
-            case serverResponse of
-                "Choose your side (Black/White):" -> do
-                    -- 让用户选择一边
-                    side <- getLine
-                    hPutStrLn handle side
-
-                "Enter the size of the board:" -> do
-                    -- 让用户输入棋盘大小
-                    size <- getLine
-                    hPutStrLn handle size
-                
-                "Send your move (format: x,y):" -> do
-                    -- 让用户输入坐标
-                    coord <- getLine
-                    hPutStrLn handle coord
-                
-                "Received" -> do
-                    -- 打印服务器确认接收坐标的消息
-                    putStrLn "Server has received the move."
-                
-                "quit" -> return ()
-
-                -- 处理棋盘状态信息
-                response | "Serialized board state:" `isPrefixOf` response -> do
-                    putStrLn $ "Server says: " ++ response
-                    let boardState = drop (length "Serialized board state: ") response
-                    putStrLn $ "Board state: " ++ boardState
-                
-                -- 处理错误信息
-                response | "Error:" `isPrefixOf` response -> do
-                    putStrLn $ "Server says: " ++ response
-                
-                -- 退出
-                -- TODO
-
-                _ -> return ()  -- 其他类型的消息不需要特殊处理
-
-            -- -- 当用户输入不是'quit'时，继续
-            -- unless (line == "quit") $ do
-            --     -- 继续循环
-            handleServerMessages handle
-        
-        waitForReceive handle = do
-            receiveMsg <- hGetLine handle
-            putStrLn $ "Server says: " ++ receiveMsg
-            unless (receiveMsg == "quit") $ do
-                handleServerMessages handle
-
+-- 从服务器获取敌方坐标
+getEnemyPawn :: Handle -> IO Coordinate
+getEnemyPawn handle = do
+    coordStr <- hGetLine handle
+    return $ parseCoordinate coordStr
 
 parseCoordinate :: String -> Coordinate
 parseCoordinate s = let [x, y] = map read $ splitOn "," s
                     in (x, y)
 
+-- 处理游戏逻辑
+-- 在 handleGame 函数中处理角色分配
+handleGame :: Handle -> IO ()
+handleGame handle = do
+    sideMsg <- hGetLine handle
+    putStrLn sideMsg -- 打印出角色信息
 
--- deserializeWholeState :: String -> WholeState
--- deserializeWholeState s = let [boardStr, playerStr, cursorStr, flagStr, winStatStr] = splitOn ";" s
---                               board = deserializeWholeBoard boardStr
---                               player = read playerStr :: PlayerSide
---                               cursor = parseCoordinate cursorStr
---                               flag = read flagStr :: Bool
---                               winStat = read winStatStr :: WinStat
---                           in WholeState board player cursor flag winStat
+    let isPlayerBlack = sideMsg == "You are Black"
+    let size = 10
+    let playCh = if isPlayerBlack then 'b' else 'w'
+    let whState = initWholeState size playCh
+
+    gameLoop handle whState isPlayerBlack
+
+-- 游戏循环
+gameLoop :: Handle -> WholeState -> Bool -> IO ()
+gameLoop handle whState isPlayer = do
+    print whState
+    if isPlayer
+    then do
+        coord <- getCoordinatesFromCommandLine
+        case updateBoardState whState coord isPlayer of
+            Left err -> do
+                putStrLn err
+                gameLoop handle whState isPlayer
+            Right newState -> do
+                sendMyPawn coord handle
+                gameLoop handle newState (not isPlayer)
+    else do
+        coord <- getEnemyPawn handle
+        case updateBoardState whState coord isPlayer of
+            Left err -> do
+                putStrLn err
+                gameLoop handle whState isPlayer
+            Right newState -> do
+                gameLoop handle newState (not isPlayer)
+
+-- 更新棋盘状态
+updateBoardState :: WholeState -> Coordinate -> Bool -> Either String WholeState
+updateBoardState state coord isPlayer = if isPlayer
+                                        then iPlacePawn state coord
+                                        else enemyPlacePawn state coord
 
 
--- deserializeWholeBoard :: String -> WholeBoard
--- deserializeWholeBoard s = let [sizeStr, boardStr] = splitOn ";" s
---                               size = read sizeStr :: Int
---                               board = map deserializeBoardDotStat $ splitOn "," boardStr
---                         in WholeBoard size [board]
-
--- deserializeBoardDotStat :: String -> BoardDotStat
--- -- deserializeBoardDotStat "0" = Left "Invalid board dot stat"
--- deserializeBoardDotStat "1" = BlackPawn
--- deserializeBoardDotStat "2" = WhitePawn
--- deserializeBoardDotStat "3" = EmptyDot
+-- 从命令行获取并解析坐标
+getCoordinatesFromCommandLine :: IO Coordinate
+getCoordinatesFromCommandLine = do
+    putStrLn "Enter your move (format: x,y):"
+    hFlush stdout  -- 确保提示信息立即打印
+    parseCoordinate <$> getLine
